20201129-3

+++ Xfs/Module/Actor/Tests/XfsActorMessageAttribute.cs
@@ -4,9 +4,9 @@ namespace Xfs
 {
 	public class XfsActorMessageAttribute : Attribute
 	{
-		public ushort Opcode { get; private set; }
+		public int Opcode { get; private set; }
 
-		public XfsActorMessageAttribute(ushort opcode)
+		public XfsActorMessageAttribute(int opcode)
 		{
 			this.Opcode = opcode;
 		}
